backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
requests
python-multipart
fastapi-cors
openai
google-generativeai